Add new fetch file and functions
